btrfs: Add Documentation/ABI/testing/sysfs-fs-btrfs file
Add the sysfs interface documentation. Here the following format is used
to record each of the sysfs btrfs entries.

What:
Date:
KernelVersion:
Contact:
Description:
Users:

Where appropriate this commit also adds, the information such as
Date, KernelVersion, Commitid if the sysfs interface has been removed
in the later kernrel.

This document is created with the aim to have clearer understanstanding
of the changes in the Btrfs sysfs interface, what are the expected
values when read.

Signed-off-by: Anand Jain <anand.jain@oracle.com>